# docs/api/swagger.yaml
definitions:
...
  server.bidResponse:
    properties:
      AgentID:
        type: integer
      Amount:
        type: string
      BuyerID:
        type: integer
      ID:
        type: integer
      ListingID:
        type: integer
      PreviousBidID:
        type: integer
    required:
    - AgentID
    - Amount
    - BuyerID
    - ID
    - ListingID
    - PreviousBidID
    type: object
  server.createBidRequest:
    properties:
      AgentID:
        type: integer
      Amount:
        type: string
      BuyerID:
        type: integer
      ListingID:
        type: integer
      PreviousBidID:
        type: integer
    required:
    - AgentID
    - Amount
    - BuyerID
    - ListingID
    type: object
...
  server.getListingByIDRequest:
    properties:
      PropertyID:
        type: integer
      Username:
        type: string
    required:
    - PropertyID
    - Username
    type: object
...
  server.listBidResponse:
    properties:
      AgentID:
        type: integer
      Amount:
        type: string
      BuyerID:
        type: integer
      ID:
        type: integer
      ListingID:
        type: integer
      PreviousBidID:
        type: integer
      Status:
        type: string
    required:
    - AgentID
    - Amount
    - BuyerID
    - ID
    - ListingID
    - PreviousBidID
    - Status
    type: object
  server.listBidsOnListingRequest:
    properties:
      ListingID:
        type: integer
    required:
    - ListingID
    type: object
  server.listBidsRequest:
    properties:
      Username:
        type: string
    required:
    - Username
    type: object
...
  server.listingResponse:
    properties:
      AcceptedBidID:
        type: integer
      AgentID:
        type: integer
      Description:
        type: string
      ID:
        type: integer
      ListingDate:
        type: string
      ListingStatus:
        type: string
      Price:
        type: string
      PropertyID:
        type: integer
    required:
    - AcceptedBidID
    - AgentID
    - Description
    - ID
    - ListingDate
    - ListingStatus
    - Price
    - PropertyID
    type: object
...
  server.rejectBidRequest:
    properties:
      ID:
        type: integer
    required:
    - ID
    type: object
...
  server.updateAcceptedBidIdByListingIdRequest:
    properties:
      AcceptedBidId:
        type: integer
      ListId:
        type: integer
    required:
    - AcceptedBidId
    - ListId
    type: object
  server.updateBidStatusRequest:
    properties:
      BidId:
        type: integer
      NewStatus:
        type: string
    required:
    - BidId
    - NewStatus
    type: object
...
info:
  contact: {}
paths:
...
  /bidding/acceptBid:
    put:
      consumes:
      - application/json
      description: accept a bid
      parameters:
      - description: accept a bid
        in: body
        name: rejectBidRequest
        required: true
        schema:
          $ref: '#/definitions/server.rejectBidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: accept a bid
      tags:
      - bidding
  /bidding/createBid:
    post:
      consumes:
      - application/json
      description: create a bid
      parameters:
      - description: create a bid
        in: body
        name: createBidRequest
        required: true
        schema:
          $ref: '#/definitions/server.createBidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.bidResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create a bid
      tags:
      - bidding
  /bidding/listBids:
    post:
      consumes:
      - application/json
      description: listing all bids belonging to a given buyer
      parameters:
      - description: listing all bids belonging to a given buyer
        in: body
        name: buyerID
        required: true
        schema:
          $ref: '#/definitions/server.listBidsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: list of bids
          schema:
            items:
              $ref: '#/definitions/server.listBidResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: given user, list all bid with them as buyer
      tags:
      - bidding
  /bidding/listLatestBidOnListing:
    post:
      consumes:
      - application/json
      description: listing most recent bid on a listing
      parameters:
      - description: listing most recent bid on a specific listing
        in: body
        name: listingID
        required: true
        schema:
          $ref: '#/definitions/server.listBidsOnListingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: bid
          schema:
            $ref: '#/definitions/server.listBidResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: given listing, list most recent bid on a listing
      tags:
      - bidding
  /bidding/rejectBid:
    put:
      consumes:
      - application/json
      description: reject a bid
      parameters:
      - description: reject a bid
        in: body
        name: rejectBidRequest
        required: true
        schema:
          $ref: '#/definitions/server.rejectBidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: reject a bid
      tags:
      - bidding
  /bidding/updateBidStatus:
    post:
      consumes:
      - application/json
      description: update a bid's status
      parameters:
      - description: update a bid status
        in: body
        name: updateBidStatusRequest
        required: true
        schema:
          $ref: '#/definitions/server.updateBidStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: update a bid's status
      tags:
      - bidding
...
  /listing/getListingByPropertyID:
    post:
      consumes:
      - application/json
      description: get listing by property id
      parameters:
      - description: get listig by property id
        in: body
        name: getListingByIDRequest
        required: true
        schema:
          $ref: '#/definitions/server.getListingByIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.listingResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get listing by property id
      tags:
      - listing
  /listing/updateAcceptedBidIdByListingId:
    post:
      consumes:
      - application/json
      description: update accepted bid id for a given listing id
      parameters:
      - description: update accepted bid id for a given listing id
        in: body
        name: updateAcceptedBidIdByListingIdRequest
        required: true
        schema:
          $ref: '#/definitions/server.updateAcceptedBidIdByListingIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: update accepted bid id for a given listing id
      tags:
      - listing
...
swagger: "2.0"